+++ src/Columns.tsx
@@ -3,10 +3,12 @@ import { SelectCellFormatter } from './formatters';
 import { Column } from './types';
 import { stopPropagation } from './utils';
 
+export const SELECT_COLUMN_KEY = 'select-row';
+
 // TODO: fix type
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const SelectColumn: Column<any, any> = {
-  key: 'select-row',
+  key: SELECT_COLUMN_KEY,
   name: '',
   width: 35,
   maxWidth: 35,
@@ -31,6 +33,19 @@ export const SelectColumn: Column<any, any> = {
         value={props.isRowSelected}
         onClick={stopPropagation}
         onChange={props.onRowSelectionChange}
+      />
+    );
+  },
+  groupFormatter(props) {
+    return (
+      <SelectCellFormatter
+        aria-label="Select Group"
+        tabIndex={-1}
+        isCellSelected={props.isCellSelected}
+        value={props.isRowSelected}
+        onChange={props.onRowSelectionChange}
+        // Stop propagation to prevent row selection
+        onClick={stopPropagation}
       />
     );
   }