+++ src/hooks/index.ts
@@ -2,4 +2,5 @@ export * from './useCombinedRefs';
 export * from './useClickOutside';
 export * from './useGridDimensions';
 export * from './useViewportColumns';
+export * from './useViewportRows';
 export * from './useFocusRef';