+++ src/types.ts
@@ -23,6 +23,8 @@ export interface Column<TRow, TSummaryRow = unknown> {
   formatter?: React.ComponentType<FormatterProps<TRow, TSummaryRow>>;
   /** Formatter to be used to render the summary cell content */
   summaryFormatter?: React.ComponentType<SummaryFormatterProps<TSummaryRow, TRow>>;
+  /** Formatter to be used to render the group cell content */
+  groupFormatter?: React.ComponentType<GroupFormatterProps<TRow, TSummaryRow>>;
   /** Enables cell editing. If set and no editor property specified, then a textinput will be used as the cell editor */
   editable?: boolean | ((row: TRow) => boolean);
   /** Determines whether column is frozen or not */
@@ -61,6 +63,7 @@ export interface CalculatedColumn<TRow, TSummaryRow = unknown> extends Column<TR
   resizable: boolean;
   sortable: boolean;
   isLastFrozenColumn?: boolean;
+  rowGroup?: boolean;
   formatter: React.ComponentType<FormatterProps<TRow, TSummaryRow>>;
 }
 
@@ -90,6 +93,17 @@ export interface FormatterProps<TRow = any, TSummaryRow = any> {
 export interface SummaryFormatterProps<TSummaryRow, TRow = any> {
   column: CalculatedColumn<TRow, TSummaryRow>;
   row: TSummaryRow;
+}
+
+export interface GroupFormatterProps<TRow, TSummaryRow = unknown> {
+  groupKey: unknown;
+  column: CalculatedColumn<TRow, TSummaryRow>;
+  childRows: readonly TRow[];
+  isExpanded: boolean;
+  isCellSelected: boolean;
+  isRowSelected: boolean;
+  onRowSelectionChange: (checked: boolean) => void;
+  toggleGroup: () => void;
 }
 
 export interface EditorProps<TValue, TRow = any, TSummaryRow = any> {
@@ -180,6 +194,20 @@ export interface RowRendererProps<TRow, TSummaryRow = unknown> extends Omit<Reac
   setDraggedOverRowIdx?: (overRowIdx: number) => void;
 }
 
+export interface GroupRowRendererProps<TRow, TSummaryRow = unknown> extends Omit<React.HTMLAttributes<HTMLDivElement>, 'style' | 'children'> {
+  id: string;
+  groupKey: unknown;
+  viewportColumns: readonly CalculatedColumn<TRow, TSummaryRow>[];
+  childRows: readonly TRow[];
+  rowIdx: number;
+  top: number;
+  level: number;
+  selectedCellIdx?: number;
+  isExpanded: boolean;
+  isRowSelected: boolean;
+  eventBus: EventBus;
+}
+
 export interface FilterRendererProps<TRow, TFilterValue = unknown, TSummaryRow = unknown> {
   column: CalculatedColumn<TRow, TSummaryRow>;
   value: TFilterValue;
@@ -212,4 +240,24 @@ export interface SelectRowEvent {
   rowIdx: number;
   checked: boolean;
   isShiftClick: boolean;
+}
+
+export type Dictionary<T> = Record<string, T>;
+
+export type GroupByDictionary<TRow> = Dictionary<{
+  childRows: readonly TRow[];
+  childGroups: readonly TRow[] | GroupByDictionary<TRow>;
+  startRowIndex: number;
+}>;
+
+export interface GroupRow<TRow> {
+  childRows: readonly TRow[];
+  id: string;
+  parentId: unknown;
+  groupKey: unknown;
+  isExpanded: boolean;
+  level: number;
+  posInSet: number;
+  setSize: number;
+  startRowIndex: number;
 }